Enable and fix the -Wshorten-64-to-32
BUG=angleproject:635

Reviewed-on: https://chromium-review.googlesource.com/316580
Tryjob-Request: Corentin Wallez <cwallez@chromium.org>
Tested-by: Corentin Wallez <cwallez@chromium.org>

+++ src/libANGLE/renderer/gl/glx/DisplayGLX.cpp
@@ -690,9 +690,9 @@ void DisplayGLX::setSwapInterval(glx::Drawable drawable, SwapControlData *data)
     }
 }
 
-bool DisplayGLX::isValidWindowVisualId(int visualId) const
+bool DisplayGLX::isValidWindowVisualId(unsigned long visualId) const
 {
-    return mRequestedVisual == -1 || mRequestedVisual == visualId;
+    return mRequestedVisual == -1 || static_cast<unsigned long>(mRequestedVisual) == visualId;
 }
 
 const FunctionsGL *DisplayGLX::getFunctionsGL() const
